breakdown timeline into smaller components
