+++ src/components/timeline/TimelineLabels.svelte
@@ -1,0 +1,9 @@
+<script>
+  export let labels = [];
+</script>
+
+<div class="timeline-labels">
+  {#each labels as label}
+    <span>{label}</span>
+  {/each}
+</div>
